remove sloshignore declarations

+++ doc/mk-site.sh
@@ -1,21 +1,21 @@
 #!/bin/bash
 
 export RUST_LOG="DEBUG"
+echo "Building docs."
 
-# clone in the legacy docs
-git clone -b gh-pages-legacy-html https://github.com/sl-sh-dev/sl-sh legacy/ || true
+if [ "${SKIP_SUPPLEMENTAL}" != "SKIP_SUPPLEMENTAL" ]; then
+    # don't want git submodules
+    rm -rf legacy/.git || true
+    mkdir slosh-rust-docs || true
+    mkdir all-rust-docs || true
 
-# don't want git submodules
-rm -rf legacy/.git || true
-mkdir slosh-rust-docs || true
-mkdir all-rust-docs || true
-
-if [ "${SKIP_DOCS}" != "SKIP_DOCS" ]; then
-    echo "Building docs."
     cargo doc --features lisp-test --target-dir all-rust-docs
     cargo doc --no-deps --document-private-items --features lisp-test --target-dir slosh-rust-docs
+
+    # clone in the legacy docs
+    git clone -b gh-pages-legacy-html https://github.com/sl-sh-dev/sl-sh legacy/ || true
 else
-    echo "Skipped doc building"
+    echo "Skipped supplemental doc building"
 fi
 
 # make the symlinks work
